Implementing reservation system, admin dashboard, code cleanup

+++ init.php
@@ -1,5 +1,7 @@
 <?php
 require_once "config.php";
+$cookieLifetime = 20 * 60;
+session_set_cookie_params($cookieLifetime);
 session_start();
 
 //============================ USER FUNCTION ============================//
@@ -19,6 +21,7 @@ function is_user_logged_in() {
 }
 
 function log_user_out() {
+        cancelCartReservation($_SESSION["user_id"]);
         unset($_SESSION["user_logged"]);
         unset($_SESSION["user_id"]);
         unset($_SESSION["user_login"]);
@@ -29,20 +32,24 @@ function log_user_out() {
 // 0 - Status: Ok
 // 1 - Status: Error while taking action, incorrect parameters / invalid action
 // 2 - Status: User not logged in
+// 3 - Status: Cart is reserved
 
 function addProductToCart($userId, $productId, $quantity) {
-    // Checking if user is logged and user_id is valid
+    // Checking if user is logged in and user_id is valid
     if(!is_user_logged_in()) return ["status" => 2, "msg" => "User not logged in."];
     // Checking if the parameters are correct
     if(!$productId || !$quantity || $quantity <= 0) return ["status" => 1, "msg" => "Incorrect product parameters."];
-
+    
     global $conn;
-
-    // Checking if the cart already exists - if not creating a new one
+    // Checking if the cart already exists - if not, creating a new one
     $cartId = getCartId($userId);
     if ($cartId === null) {
         $cartId = createCart($userId);
     }
+
+    // Checking if cart is reserved
+    if(checkCartReservation($userId)) return ["status" => 3, "msg" => "Cannot edit cart during reservation."];
+
     // Checking if the product is already in the cart
     $sql = "SELECT quantity FROM cart_item WHERE cart_id = ? AND product_id = ?";
     $stmt = $conn->prepare($sql);
@@ -54,8 +61,6 @@ function addProductToCart($userId, $productId, $quantity) {
         $row = $result->fetch_assoc();
         if ($row) {
             $currentQuantityInCart = $row["quantity"];
-        } else {
-            $currentQuantityInCart = 0;
         }
     }
 
@@ -68,6 +73,11 @@ function addProductToCart($userId, $productId, $quantity) {
 
     if ($inventoryRow = $inventoryResult->fetch_assoc()) {
         $availableQuantity = $inventoryRow["product_quantity"];
+
+        // Check if product quantity in inventory is 0, if so call removeProductFromCart
+        if ($availableQuantity == 0) {
+            return removeProductFromCart($userId, $productId);
+        }
 
         $totalQuantity = $currentQuantityInCart + $quantity;
         if ($totalQuantity > $availableQuantity) {
@@ -101,9 +111,10 @@ function addProductToCart($userId, $productId, $quantity) {
     }
 }
 
+
 function subtractProductFromCart($userId, $productId, $quantity) {
     // Checking if user is logged and user_id is valid
-    if(!is_user_logged_in()) return ["status" => 2, "msg" => "User not logged in."] ;
+    if(!is_user_logged_in()) return ["status" => 2, "msg" => "User not logged in."];
     // Checking if the parameters are correct
     if(!$productId || !$quantity || $quantity <= 0) return ["status" => 1, "msg" => "Incorrect product parameters."];
 
@@ -115,6 +126,22 @@ function subtractProductFromCart($userId, $productId, $quantity) {
         return ["status" => 1, "msg" => "Cart not found."];
     }
 
+    // Checking if cart is reserved
+    if(checkCartReservation($userId)) return ["status" => 3, "msg" => "Cannot edit cart during reservation."];
+
+    // Checking inventory quantity
+    $inventorySql = "SELECT product_quantity FROM Inventory WHERE product_id = ?";
+    $inventoryStmt = $conn->prepare($inventorySql);
+    $inventoryStmt->bind_param("i", $productId);
+    $inventoryStmt->execute();
+    $inventoryResult = $inventoryStmt->get_result();
+
+    if ($inventoryRow = $inventoryResult->fetch_assoc()) {
+        if ($inventoryRow["product_quantity"] == 0) {
+            return removeProductFromCart($userId, $productId);
+        }
+    }
+
     // Fetching the product quantity from cart_item
     $sql = "SELECT quantity FROM cart_item WHERE cart_id = ? AND product_id = ?";
     $stmt = $conn->prepare($sql);
@@ -122,7 +149,7 @@ function subtractProductFromCart($userId, $productId, $quantity) {
     $stmt->execute();
     $cartResult = $stmt->get_result();
 
-    if($row = $cartResult->fetch_assoc()) {
+    if ($row = $cartResult->fetch_assoc()) {
         $currentQuantityInCart = $row["quantity"];
         $newQuantity = $currentQuantityInCart - $quantity;
 
@@ -139,7 +166,7 @@ function subtractProductFromCart($userId, $productId, $quantity) {
             } else {
                 return ["status" => 1, "msg" => "Error updating product quantity."];
             }
-         }
+        }
     } else {
         return ["status" => 1, "msg" => "Product not found in cart."];
     }
@@ -159,6 +186,9 @@ function removeProductFromCart($userId, $productId) {
         return ["status" => 1, "msg" => "Cart not found."];
     }
 
+    // Checking if cart is reserved
+    if(checkCartReservation($userId)) return ["status" => 3, "msg" => "Cannot edit cart during reservation."];
+
     // Deleting product from cart
     $sql = "DELETE FROM cart_item WHERE cart_id = ? AND product_id = ?";
     $stmt = $conn->prepare($sql);
@@ -172,8 +202,8 @@ function removeProductFromCart($userId, $productId) {
 }
 
 function updateProductFromCart($userId, $productId, $quantity) {
-    // Checking if user is logged and user_id is valid
-    if(!is_user_logged_in()) return ["status" => 2, "msg" => "User not logged in."] ;
+    // Checking if user is logged in and user_id is valid
+    if(!is_user_logged_in()) return ["status" => 2, "msg" => "User not logged in."];
     // Checking if the parameters are correct
     if(!$productId || $quantity < 0) return ["status" => 1, "msg" => "Incorrect product parameters."];
     
@@ -181,9 +211,12 @@ function updateProductFromCart($userId, $productId, $quantity) {
     
     // Fetching cart value
     $cartId = getCartId($userId);
-     if ($cartId === null) {
+    if ($cartId === null) {
         return ["status" => 1, "msg" => "Cart not found."];
     }
+
+    // Checking if cart is reserved
+    if(checkCartReservation($userId)) return ["status" => 3, "msg" => "Cannot edit cart during reservation."];
 
     // Fetching the product quantity from Inventory
     $sql = "SELECT product_quantity FROM inventory WHERE product_id = ?";
@@ -195,10 +228,11 @@ function updateProductFromCart($userId, $productId, $quantity) {
     if ($inventoryRow = $inventoryResult->fetch_assoc()) {
         $availableQuantity = $inventoryRow["product_quantity"];
 
-        // Checking if updated quantity is equal or below 0
-        if ($quantity == 0) {
+        // Check if product quantity in inventory is 0, if so call removeProductFromCart
+        if ($quantity == 0 || $availableQuantity == 0) {
             return removeProductFromCart($userId, $productId);
         }
+
         // Checking if updated quantity is above quantity in inventory
         if ($quantity > $availableQuantity) {
             if ($availableQuantity > 0) {
@@ -207,7 +241,7 @@ function updateProductFromCart($userId, $productId, $quantity) {
                 return ["status" => 1, "msg" => "Product is out of stock."];
             }
         }
-        // Upadting product quantity
+        // Updating product quantity
         $sql = "UPDATE cart_item SET quantity = ? WHERE cart_id = ? AND product_id = ?";
         $stmt = $conn->prepare($sql);
         $stmt->bind_param("iii", $quantity, $cartId, $productId);
@@ -220,6 +254,7 @@ function updateProductFromCart($userId, $productId, $quantity) {
         return ["status" => 1, "msg" => "Product not found in inventory."];
     }
 }
+
 
 // Function that checks if the cart was already created for the user
 function getCartId($userId) {
@@ -292,7 +327,7 @@ $userFirstName, $userLastName, $userEmail, $userPhoneNumber, $userHomeAddress, $
         return ["status" => 1, "msg" => "Incorrect address parameters"];
     }
     //return ["status" => 1, "msg" => htmlspecialchars($userFirstName). " " .htmlspecialchars($userLastName). " " .htmlspecialchars($userEmail). " " .htmlspecialchars($userPhoneNumber). " " .htmlspecialchars($userHomeAddress). " " .htmlspecialchars($userCity). " " .htmlspecialchars($userPostalCode). " " .htmlspecialchars($userState). " " .htmlspecialchars($userCountry). " "];
-    
+
     global $conn;
     $sql = "CALL AddUserAddress(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, @p_status, @p_message)";
     $stmt = $conn->prepare($sql);
@@ -360,19 +395,65 @@ function deleteUserAddress($userId, $userAddressId) {
     } 
 }
 
-//============================ ORDER FUNCTION ============================//
-// 0 - Status: Ok
-// 1 - Status: Error while taking action, incorrect parameters / invalid action
-// 2 - Status: User not logged in
-
+//============================ RESERVATION / ORDER FUNCTION ============================//
 function startCartReservation($userId) {
     // Checking if user is logged and user_id is valid
-    if(!is_user_logged_in() || !isCartExist($userId)) return null;
+    if (!is_user_logged_in() || !isCartExist($userId)) return;
 
-
+    global $conn;
+    $sql = "CALL StartCartReservation(?)";
+    $stmt = $conn->prepare($sql);
+    $stmt->bind_param("i", $userId);
+    $stmt->execute();
 }
 
+function checkCartReservation($userId) {
+    // Checking if user is logged and user_id is valid
+    if (!is_user_logged_in() || !isCartExist($userId)) return false;
 
+    global $conn;
+    $sql = "SELECT cart_reserved, cart_reservation_time FROM cart WHERE user_id = ?";
+    $stmt = $conn->prepare($sql);
+    $stmt->bind_param("i", $userId);
+    $stmt->execute();
 
+    $result = $stmt->get_result();
+    if ($row = $result->fetch_assoc()) {
+        return $row['cart_reserved'] == 1 && !is_null($row['cart_reservation_time']);
+    }
+    return false;
+}
+
+function cancelCartReservation($userId) {
+    // Checking if user is logged and user_id is valid
+    if (!is_user_logged_in() || !isCartExist($userId)) return;
+
+    global $conn;
+    $sql = "CALL CancelCartReservation(?)";
+    $stmt = $conn->prepare($sql);
+    $stmt->bind_param("i", $userId);
+    $stmt->execute();
+}
+
+function finalizeOrder($userId, $userFirstName, $userLastName, $userEmail, $userPhoneNumber, $userHomeAddress, $userCity, $userPostalCode, $userState, $userCountry, $userDeliveryMethod, $userPaymentMethod) {
+    // Checking if user is logged and cart are valid and exist
+    if (!is_user_logged_in()) return ["status" => 2, "msg" => "User not logged in."];
+    if (!isCartExist($userId)) return ["status" => 1, "msg" => "Cart not found."];
+
+    global $conn;
+    $sql = "CALL CreateOrder(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, @p_status, @p_message)";
+    $stmt = $conn->prepare($sql);
+    $stmt->bind_param("isssssssssss", $userId, $userFirstName, $userLastName, $userEmail, $userPhoneNumber, $userHomeAddress, $userCity, $userPostalCode, $userState, $userCountry, $userDeliveryMethod, $userPaymentMethod);
+    $stmt->execute();
+    $select = $conn->query("SELECT @p_status AS status, @p_message AS message");
+    $result = $select->fetch_assoc();
+
+    // Check if the order was created
+    if($result) {
+        return ["status" => $result['status'], "msg" => $result['message']];
+    } else {
+        return ["status" => 1, "msg" => "Error while creating order."];
+    }
+}
 ?>
